refactor: add PeerPoolHandling::get_best_connected_peers
